validation

+++ SWOMT/Views/Certificats.xaml.cs
@@ -22,13 +22,38 @@ namespace SWOMT.Views
     public partial class Certificats : Page
     {
         List<MyApps.Application.ViewModels.CertificatViewModel> liste = new List<MyApps.Application.ViewModels.CertificatViewModel>();
+        List<MyApps.Application.ViewModels.ParticipantViewModel> listeParticipant = new List<MyApps.Application.ViewModels.ParticipantViewModel>();
+        List<MyApps.Application.ViewModels.ResultatViewModel> ListeModulesEchoues = new List<MyApps.Application.ViewModels.ResultatViewModel>();
+        List<MyApps.Application.ViewModels.ResultatViewModel> ListeModulesReussis = new List<MyApps.Application.ViewModels.ResultatViewModel>();
+
 
         string enregistre;
-        public Certificats()
+       
+        public Certificats(string roleName)
         {
             InitializeComponent();
             liste = MyApps.Application.Services.CertificatViewModelService.GetCertificats();
-            PopulateAndBind(liste);
+         
+            listeParticipant = MyApps.Application.Services.ParticipantsViewModelServices.GetParticipants();
+            PopulateAndBindParticipant(listeParticipant);
+            ListeModulesEchoues = MyApps.Application.Services.ResultatsVieModelService.GetResultats();
+            ListeModulesReussis = MyApps.Application.Services.ResultatsVieModelService.GetResultats();
+            if ((string)roleName != "Admin")
+            {
+
+               
+                // FormateurtextBox.Content=false;
+                CertificatTextBox.IsEnabled = false;
+
+            }
+            if ((string)roleName != "Secrétaire" && (string)roleName != "Admin" && (string)roleName != "Formateur")
+            {
+
+                IdNationale.Visibility = Visibility.Hidden;
+                Datenaissance.Visibility = Visibility.Hidden;
+                email.Visibility = Visibility.Hidden;
+                telephone.Visibility = Visibility.Hidden;
+            }
 
         }
 
@@ -62,16 +87,21 @@ namespace SWOMT.Views
                 IdCertificat.Text = donnee.IdCertificat.ToString();
                 IdParticipant.Text = donnee.IdParticipant.ToString();
                 NomParticipant.Text = donnee.NomParticipant.ToString();
-                DateDelivrance.Text = donnee.DateDelivrance.ToString();
+                //DateDelivrance.Text = donnee.DateDelivrance.ToString();
 
             }
+           
         }
 
         private void Ajouter_Click(object sender, RoutedEventArgs e)
         {
             enregistre = "Ajouter";
-            ClearFormValues();
-            ModeIsEnabledTrue();
+            
+            ModeIsEnabledFalse();
+            
+            NomParticipant.Text = "";
+            DateDelivrance.Text = "";
+            DateDelivrance.IsEnabled = true;
         }
         /// <summary>
         /// méthode pour mettre à jour et ajouter  une site
@@ -80,21 +110,31 @@ namespace SWOMT.Views
         /// <param name="e"></param>
         private void MettreAjour_Click(object sender, RoutedEventArgs e)
         {
-            Certificat element = new Certificat(); 
+            Certificat element = new Certificat();
             //Competence competence = new Competence();
 
-            //if (NomCompetence.Text == "")
-            //{
-            //    MessageBox.Show("Il faut saisir le nom de la competence");
-            //    return;
-            //}
+            if (IdParticipant.Text == "")
+            {
+                MessageBox.Show("Il faut sélectionner un élément dans liste des certificats délivrés ");
+                return;
+            }
+            if (NomParticipant.Text == "")
+            {
+                MessageBox.Show("Il faut sélectionner un participant réussi un module ");
+                return;
+            }
 
+            if (DateDelivrance.Text=="")
+            {
+                MessageBox.Show("Il faut sélectionner dans la liste des modules réussis par un participant");
+                return;
+            }
 
 
             if (enregistre == "Ajouter")
             {
                 element.IdParticipant = short.Parse(IdParticipant.Text);
-                element.DateDelivrance = DateTime.Parse(DateDelivrance.Text).Date;
+                element.DateDelivrance = DateTime.Now;
 
                 MyApps.Domain.Service.CertificatService.Create(element); 
 
@@ -109,12 +149,13 @@ namespace SWOMT.Views
 
                 MyApps.Domain.Service.CertificatService.Update(element);
             }
-
-
-            ModeIsEnabledFalse();
             liste.Clear();
-            liste = MyApps.Application.Services.CertificatViewModelService.GetCertificats();
+            liste = MyApps.Application.Services.CertificatViewModelService.GetCertificatsPerParticipant(short.Parse(IdParticipant.Text));
+            NbrCertificatRécu.Text = liste.Count().ToString();
             PopulateAndBind(liste);
+            DateDelivrance.Text = ""; 
+            ModeIsEnabledFalse();
+           
 
         }
         /// <summary>
@@ -131,7 +172,7 @@ namespace SWOMT.Views
                 return;
             }
             enregistre = "Modifier";
-            ModeIsEnabledTrue();
+            ModeIsEnabledFalse();
 
 
         }
@@ -143,15 +184,22 @@ namespace SWOMT.Views
         /// <param name="e"></param>
         private void Supprimer_Click(object sender, RoutedEventArgs e)
         {
-            //le code pour signaler la presence de l'idParticipant dans la table Inscription on doit d'abord faire une vérification
+            //tester si l'identifiant n'est pas null 
+            if (IdCertificat.Text == "")
+            {
+                MessageBox.Show("Il faut saisir identifiant de certificat ");
+                return;
+            }
 
             MyApps.Domain.Service.CertificatService.Delete(short.Parse(IdCertificat.Text));
 
-            ClearFormValues();
+          
+            
             liste.Clear();
-            liste = MyApps.Application.Services.CertificatViewModelService.GetCertificats();
+            liste = MyApps.Application.Services.CertificatViewModelService.GetCertificatsPerParticipant(short.Parse(IdParticipant.Text));
+            NbrCertificatRécu.Text = liste.Count().ToString();
             PopulateAndBind(liste);
-
+            ClearFormValues();
         }
         private void ClearFormValues()
         {
@@ -177,5 +225,136 @@ namespace SWOMT.Views
             DateDelivrance.IsEnabled = false; 
 
         }
+        //************************************************************************************************************************************************
+        //************************************** LA LISTE DES PARTICIPANT PARTIE DE DROITE ***************************************************************
+
+        private void PopulateAndBindParticipant(List<MyApps.Application.ViewModels.ParticipantViewModel> listeCompetences)
+        {
+            Binding monBinding = new Binding
+            {
+                Path = new PropertyPath("Value")
+            };
+           
+            ListParticipant.DataContext = listeCompetences;
+
+        }
+        private void ListParticipant_MouseDoubleClick(object sender, SelectionChangedEventArgs e)
+        {
+            if (ListParticipant.SelectedItem is MyApps.Application.ViewModels.ParticipantViewModel donnee)
+            {
+
+                IdParticipant.Text = donnee.IdParticipant.ToString();
+                ListeModulesReussis = MyApps.Application.Services.ResultatsVieModelService.GetListModulesRéussis(short.Parse(IdParticipant.Text));
+                //ListeModulesReussis = MyApps.Application.Services.ResultatsVieModelService.GetListModulesRéussis(
+                //     MyApps.Domain.Service.ResultatService.GetIdInscription(donnee.IdParticipant));
+                PopulateAndBindRéussis(ListeModulesReussis);
+                TotalRéussi.Text = ListeModulesReussis.Count().ToString();
+
+                ListeModulesEchoues = MyApps.Application.Services.ResultatsVieModelService.GetListModulesEchoué(short.Parse(IdParticipant.Text));
+                PopulateAndBindParticipantFailed(ListeModulesEchoues);
+                TotalEchoué.Text = ListeModulesEchoues.Count().ToString(); 
+
+                IdParticipant.Text = donnee.IdParticipant.ToString();
+                NomParticipant.Text = donnee.NomParticipant.ToString();
+                liste.Clear();
+                liste = MyApps.Application.Services.CertificatViewModelService.GetCertificatsPerParticipant(short.Parse(IdParticipant.Text));
+                NbrCertificatRécu.Text = liste.Count().ToString();
+                PopulateAndBind(liste);
+            }
+
+           
+
+        }
+        private void Rechercher_Click(object sender, RoutedEventArgs e)
+        {
+
+            if (NomRechercher.Text == "")
+            {
+                MessageBox.Show("Entrer le nom à rechercher");
+                listeParticipant = MyApps.Application.Services.ParticipantsViewModelServices.GetParticipants();
+                PopulateAndBindParticipant(listeParticipant);
+                return;
+
+            }
+
+            listeParticipant = MyApps.Application.Services.ParticipantsViewModelServices.GetParticipantByMethodeSearch(NomRechercher.Text);
+            PopulateAndBindParticipant(listeParticipant);
+        }
+        private void ReSetList_Click(object sender, RoutedEventArgs e)
+        {
+            NomRechercher.Text = "";
+            listeParticipant = MyApps.Application.Services.ParticipantsViewModelServices.GetParticipantByMethodeSearch(NomRechercher.Text);
+            PopulateAndBindParticipant(listeParticipant);
+        }
+
+        //************************************************************************************************************************************
+        //************************************* Partie à gauche : les liste des participant réussis et échoués********************************
+
+      
+        
+        private void PopulateAndBindRéussis(List<MyApps.Application.ViewModels.ResultatViewModel> listeItems)
+        {
+            Binding monBinding = new Binding
+            {
+                Path = new PropertyPath("Value")
+            };
+            ListParticipantRéussi.DataContext = listeItems;
+        }
+
+        /// <summary>
+        /// binding la liste des participant echoués
+        /// </summary>
+        /// <param name="listeItems"></param>
+        private void PopulateAndBindParticipantFailed(List<MyApps.Application.ViewModels.ResultatViewModel> listeItems)
+        {
+            Binding monBinding = new Binding
+            {
+                Path = new PropertyPath("Value")
+            };
+            ListParticipantFailed.DataContext = listeItems;
+        }
+        private void ListParticipantRéussi_MouseDoubleClick(object sender, SelectionChangedEventArgs e)
+        {
+
+            if (ListParticipantRéussi.SelectedItem is MyApps.Application.ViewModels.ResultatViewModel donnee)
+            {
+                if (donnee.IdExamen == 0)
+                {
+                    MessageBox.Show("Selectionner un élément dans la liste");
+                    return;
+                 
+                }
+                enregistre = "Ajouter";
+                NomParticipant.Text = donnee.NomParticipant.ToString();
+                DateDelivrance.Text = DateTime.Now.ToString();
+
+
+
+            }
+
+
+        }
+
+        private void ListParticipantFailed_MouseDoubleClick(object sender, SelectionChangedEventArgs e)
+        {
+
+            if (ListParticipantFailed.SelectedItem is MyApps.Application.ViewModels.ResultatViewModel donnee)
+            {
+                if (donnee.IdExamen == 0)
+                {
+                    MessageBox.Show("Selectionner un élément dans la liste");
+                    return;
+                 
+                }
+               
+            }
+          
+        }
+        private void ComboBoxModulesPerParticipant_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            //code pour faire une rechercher dans la liste des participants réussis 
+
+        }
+
     }
 }
